Tests/LibWeb/Text/expected/css/style-ffi-counters.txt

has counters: true
all zero after reset: true
longhand driver ran: true
longhand loop called back into C++: true
